use smaller buffers to read files

+++ src/main/kotlin/com/github/hoxo/fileservice/FileService.kt
@@ -6,7 +6,7 @@ import java.nio.file.Path
 interface FileService {
     suspend fun getInfo(path: String): Result<FileInfo>
     suspend fun list(path: String): Result<Flow<FileInfo>>
-    suspend fun readFile(path: String, offset: Long, toRead: Long): Result<ByteArray>
+    suspend fun readFile(path: String, offset: Int, toRead: Int): Result<ByteArray>
     suspend fun createEmptyFile(path: String): Result<FileInfo>
     suspend fun createEmptyDir(path: String): Result<FileInfo>
     suspend fun delete(path: String): Result<Unit>
